navigation: catch errors on section config's functions
fixes #1027

Each user-defined functions in the `SectionConfig` type is wrapped
around a try/catch, which allows to give a complete message on which
function/section the error occurred in. The error is thrown and should
be caught by the survey's catchAll mechanism and send to the server.

Add a `sectionName` property to the `SectinConfigWithDefaults*` types to
more easily identify the name of the section without having to pass it
around in the navigation service.