inputs: let the checkbox and multiselect parse arrays
fixes #1036

Add a `asArray` parameter to the `parseValue` function, which allows to
specify if the expected value should be an array, but still convert to
the correct datatype the values in the array.

Add a `isInputTypeWithArrayValue` function to determine if the input
type should give an array value. checkboxes and multiselects are inputs
with array values.