survey: add `languageChange` and `interviewOpen` user actions
This adds a migration to add the `language_change` and `interview_open`
types to the event type in the paradata event table.

Type new `UserAction`s with `languageChange` and `interviewOpen` as
types and respectively the new and previous language, and the browser
and requested language.

Unit test all the changes to the paradata export and logging.