backend: split user action side effect from real effect
Some user actions, for example `sectionChange` and eventually
`languageChange` and `interviewOpen` have side effects on the response
and need to update some interview response that may map to response
or corrected_response, so these changes need to be added before the mapping.
Note that these side-effects should be mandatory until the interview
structure is extracted from the json field.

The real effects, ie the changes from the user actions that are
standalone and complete in themselves, like the widget interaction that
updates a specific response path, are handled directly by the update.