survey: support captcha in support form and byField
fixes #1021

The support form always provides a captcha. The request form is not
expected to give any error and should be sent only once to the server,
so the captcha token is not kept between requests and if the request
fails, the captcha needs to be answered again.

As for `byField` login form, it will display a captcha after the first
un-successful login attempts. Since typos may happen and there may be
more than one attempt at login with access code and postal code
combination, the captcha is kept between requests and the user does not
need to answer it again.

The backend route for support and byField login have been updated to use
the captcha validation middleware.