generator: let radioNumber widgets support a field for min/max
fixes #1050

If the min/max value entered is not an integer, it is considered a field
value and the min/max value will be replaced by a function that gets the
corresponding response.

The `generate_widget_statement` function now returns not just the
statement, but optional values that hints at additional imports to add
to the import statement. This allows to add the `surveyHelper` functions
import from the `evolution-common/lib/utils/helpers` file. If the
radioNumber widget uses the `surveyHelper` function in its code, the
`generate_widget_statement` function will return `True` for
`needsHelperImport`.