uiTests: add helper for login with access/postal codes
Extract the landing page test sequence from the register functions, as
the landing page is always the same in all tests.

Add a test helper to register with access code and postal code. This
helper takes an access code, a postal and an optional
`isExpectedToExist` boolean value if the login is expected to be done in
one go or if the user should check the warning checkbox.