formatters: add formatter functions for access and postal codes
Add a dependency to the `cleave-zen` library which contains formatting
functions for various use cases, like blocks with delimiters.

Add a file in `evolution-common/lib/utils`, called `formatters.ts` that
will contain formatter functions.

Add a `eightDigitsAccessCodeFormatter` which formats a string in the XXXX-XXXX
format where X are digits.

Add a `canadianPostalCodeFormatter` function which formats a string as
an uppercase `XXX XXX` format where X are digits or letters.