authByField: use the formatters from the survey fields
fixes #1053

To avoid the `ByFieldLoginForm` to hard-code formats of the fields, we
directly use fields of the survey that are defined in the widgets and use
the `inputFilter`, if any, for the corresponding field input in the auth
form.

The validations from the input are also used for the login form. Since
those validations are called in a context that does not have a valid
interview yet, the validation functions of those fields should be simple
and not depend on other answers. There may be unexpected behavior if it
is not the case.

This updates the configuration of the `byField` auth method. Instead of
a simple `true` or `false` value, it can be an object to configure the
name of the field to use for access code (`accessCodeField`) or postal
code (`postalCodeField`). By default, the auth form will look for fields
named `accessCode` and `postalCode`.