Fix bugs with the generate_labels function with multiple sheet.
Correct a bug that removes existing labels during the wrong operation.
Fix a bug that generates multiple headers in labels.
Remove function delete_yaml_file that is no longer used, because we use delete_all_labels_yaml_files.
Remove removed_files_global that was no longer used and the tests with it.