generator: support inputFilter parameter for string widgets
The `parameters` field for a `String` widget can add a value of
`formatter=nameOfFormatter`. The name of the formatter can be a function
that is in the `formatters.ts` file in `evolution-common/src/utils`. It
can also be a custom formatter. If it ends with `CustomFormatter`, it
will add an import to a `customFormatters.ts` file that should be
present in the generated survey.

Add simple tests for the string widget generation.

Also rename the `needsHelperImport` to `has_helper_import`