formatters: use `delimiterLazyShow` in block formatting
If not, then if the user erases the formatted string, it is not possible
to erase the delimiter, it will keep coming back, which implies the rest
of the string cannot be erased either.

With `delimiterLazyShow`, the user does not have to enter the delimiter,
it will always automatically be set as the rest of the string is
entered. But it implies that the delimiter will not be shown if the user
enters it, but it's better than not being able to remove the string.