widgetOperation: fix previous status when path is changed
fixes #1116

When a previous status exists for a widget name, but it does not have
the same path as the one of the widget being prepared, do not return the
previous status. This extracts the previous status fetch to a function
which does this check.

Add a unit test the validates this change.