generator: add a fourth option for custom vs other gender string
fixes #1111

This allows to add another gender option, that will be used as the
default, for more neutral tones (like if the participant prefers not to
answer), while a custom option is when the gender is some other, but
specific, custom option.

The gender string now supports four parts: male/female/custom/other. If
there are only 2 parts, custom and other will use the first one (same as
male). If there are three parts, other and custom will be the same.

Also add support for quoted gender strings, which allow to have a '/'
character inside the quotes. For example
`{{gender:il/elle/iel/"il/elle"}}` would resolve in 4 parts,
respectively, "il", "elle", "iel" and "il/elle".

Add tests for all those cases.