icons: add an activity to icon mapping
Update Transition submodules to get the marker icons.

Add a map of activities and activity categories to the path of the
corresponding icon. As there is no exhaustive list of possible
activities, the activities were taken from the od_nationale_quebec. It
maps to the best possible icon representing the activity.

Also add a function to retrieve the icon for a specific activity or
category, which returns the default icon if the activity mapping does not
exist. Another function returns a map marker for the activity and activity
category.

It implies that for each activity, there is a basic icon for the activity
with a `.svg` extension, but also marker icons, ie icons with a pointy
end to mark a location, with suffix `-marker_round.svg` and
`-marker_square.svg`, to support either a round or square marker.