ui tests: support a regular expression for popup text
The popup text may contain placeholder, like the name of the
participant. It should be possible to specify a regular expression for
the `popupText` of the `inputPopupButtonTest`.

Example usage: if a popup has a text with `Hello {{nickname}}`, we can
send `/Hello .*/` for the popupText value instead of a string `"Hello I
don't know what to put here"`.