helper: fix return value when `undefined` value and `geojson`
fixes #1130

If the value passed to the `parseValue` helper function is `undefined`,
that value should be returned as well, not `null`. All other types had
unit tests to make sure this case was covered, except the geojson and it
ended up buggy.

This accidentally caused a widget to be not responded while it should
be, which explains the seemingly unrelated issue that it fixes.