generator: use `gender` and/or `sexAssignedAtBirth` for label context
fixes #1143

This checks if a `gender` or `sexAssignedAtBirth` field exist in the
survey and it will use either in the context if the `{{gender:` is
present in the label. If both fields are present, `gender` has
precedence. If none are present, it issues a warning. The fields are
checked at the beginning of the section, so using a gendered label in
a section before the field is defined will issue a warning.