validations: support Canada or Quebec postal code validations
fixes #1150

This extracts 2 regex for Canada and Quebec postal codes. The survey can
be configured to use one of those formats and everywhere it is needed
(postal code validation, authentication by access/postal codes) will use
this configuration.

Add a `postalCodeRegion` configuration, which can take the value of
either `quebec` or `canada`. If undefined, the postal code's content
will not be validated with the builtin functions. Surveys can add their
own validation functions if need be.