generator: add gendered suffix labels if gendered already exist
fixes #1160

The translation strings use the gender as context and the plural suffix to
address participant directly or with a nickname. In i18n, the
context has priority over the plural, so, for example, with both, the
labels will be looked for in priority: `label_context_suffix`,
`label_context`, `label_suffix`, `label`. In Evolution, the plural (suffix)
should have precedence over the context, so we need to make sure that if a
label with `label_context` exists and there is plural, then
`label_context_suffix` should also exist, otherwise, we'll get the
multi-person gendered translation before the single-person one.