backend: escape html from string values
part of #1169

Add depencies in the backend to `dompurify` and `jsdom`

Whenever the interview is updated, all values by path of type strings,
even lower in objects, are sanitized, to avoid javascript injection and
XSS attacks.

Those values need to be sent back to the client in their sanitized form,
so are added to the server updated fields, only if the server did not
himself update that field.