generator: escape the nickname in labels
fixes #1169

Since the nickname comes from the user, to avoid XSS attacks or
javascript injection in the code, escape the string with
`lodash/escape`.

** Breaking change **

The package should be added to dependencies of the surveys using
the generator if it has labels with nickname. Note that the strings
should have been sanitized on the server already so this is just an
added safeguard to our labels.