input: keep cursor position when inputFilter applied
fixes #1180

When the inputFilter function returns a new string object (ie does not
mutate in place the current one), then the `input` element sees this new
value and some browser simply bring the cursor at the end of this new
string.

We add a `FormattedInput` component, which wraps the input to save the
cursor position before formatting the value and attempts to return to
this position after the update. The formatted value differs from the
input value so we attempt a best guess, ie keeps the previous position
if the string has the same length or, if formatted string is longer,
advances by the difference if the characters after the previous position
are the same as the end of the formatted string, or keep the current
position otherwise.

Use this `FormattedInput` component in the `InputString` component if
the `inputFilter` is set (to avoid the overhead on normal text fields),
and in the `ByFieldLoginForm`, on the login components which have a
formatter for access code and postal code.