formatters: do not use cleave-zen for postal code formatting
The cleave-zend block formatter was either too agressive or too lazy
with the delimiter, resulting in the user not being able to erase it (if
`delimiterLazyShow` is set to `true`), or not being able to write it (if
`false`). We use a custom uppercase formatter which better supports the
space delimiter, adding it automatically when all characters are
entered.