authByField: add a random string to username when mismatch
fixes #1178

If the accessCode and postalCode do not match a pair that is in the
prefilled data, we add a 6 hex digits random string to the username.
This will prevent leaking personal data if someone entered some invalid
and easily guessed values, like 1234-1234 and 'H1H 1H1' and filled the
survey.