survey: fix race condition in `startUpdateInterview` redux action
fixes #1187

With slower connections, many concurrent update operations can be run in
parallel and the result of the subsequent will override the previous
results. Update operations need to be run sequentially.

We used to have the `redux-async-queue` package to do this, but it does
not work with latest version of redux that we used. It is not supposed
to be necessary, according to redux documentation, but they also
strongly suggest using `@reduxjs/toolkit`, which we have not done yet.

In the meantime, we need to run update in sequence, so we enqueue them
in a 1-concurrent p-queue. We can investigate later if our use of redux
is wrong, too complex or if the a new package would solve the problem.

Add a test for concurrent update calls. This test failed with previous
code, but passes with this commit.