infoMap: make sure the `properties` of the geography are defined
Some errors in survey may make the `properties` field absent in
geographies. If it is the case, the survey will just become impossible
to open with an unrecoverable error. This adds a check to make sure the
`properties` exist, with a console warning if it doesn't so we know
there's something to fix.