segments: fix trip sequence when they are created
fixes #1206

Trips were added one at a time, but they were not directly added to the
interview, which means the number of trips did not change, so all trips
were created with the same sequence number. Instead, we keep an array of
trips data to add and create all trips at once, such that they will each
have a different sequence number.

Also do not mock the `addGroupedObject` function to allow to catch such
errors as the one we had.