survey: unset undefined valuesByPath instead of setting undefined
fixes #1216

When updating the interviews, the keys in `valuesByPath` that map to an
`undefined` value should be unset instead of being set to `undefined`,
which effectively kept the key present in the object and caused errors
in group objects like trips or visited places, which were still counted
as existing.

An `undefined` value in the `valuesByPath` is the equivalent to setting
it in the `unsetPaths`, as this is what is done before sending the
values to the server, such that the interview, on the server side will
effectively unset this value.

Also update the unit tests to use the deepEqual on the interview, as
the jest equality will consider equal an undefined value for a key and
the absence of the key. So it makes sure the interview as really as
expected.