ui: fix scroll to invalid text field in iOS devices
fixes #1248

Some iOS browser have a security restrictions to prevent
programmatically focusing on elements without direct user interaction.
This preventing from scrolling to the invalid field if that field was a
text field.

Scroll to the first invalid field in all cases, and try to focus on text
field. If it does not work, at least the invalid field will be visible
on screen.

Change scrolling method from `window.scrollTo` to
`element.scrollIntoView` as the latter focuses more on the element
itself and should be preferred when the intention is to make the element
visible.