Properly handle language in participant interview
And ignore the reviewer's

fixes #1251

Only the participant's interview tracks language changes in the
paradata, so the `useEffect` hook is moved from the `Survey` component
to `SurveyParticipant`. This will affect both respondent and
interviewer. Reviewer's language changes will be ignored.

Set the language when the interview is activated, ie when first
activating an interview, or when the interview is opened again, but the
language is different from the previous one, the language of the
interview will be updated to the current interface's.