eslint: updater rule arguments for `@typescript-eslint/no-unused-vars`
Add `"varsIgnorePattern": "^_"` to let the linter know that variables
preceded with `_` are known to be unused. The previous options only
allowed this for function arguments, but not variables. This will allow
to use destructuring to remove some field from an object, for example
`const { response: _response, ...rest } = interview` used to give linter
error on the _response field, but we do want to remove it from the
interview object and it should be correct.