admin: add warnings for resetting or editing interviews
part of #1257

The button to reset interviews to participant's data now has a warning
to confirm the action, to avoid accidental clicks on this button.

The button to edit the interview now gives a warning if the interview
was already reviewed/flagged. It does not prevent from editing, it just
warns that a decision was already made on that interview and that
editing it may cause incoherences.