calculations: add accessibility map calculation function
And implement it using the Transition API.

Add types for the query parameters and results. This returns the geojson
polygons, so it makes the query to fetch the geojson polygons as well. Not all
parameters are implemented for now. It supports the `numberOfPolygons`
and `calculatePois` parameters, as well as the total travel time.